Revert "Fix cross-worker revalidate API (#49101)"
This reverts commit 931c10129913a847cf6c4ba5f561b2ebf5b8edcc.
