--- test/production/app-dir/revalidate/app/layout.js
@@ -1,7 +1,0 @@
-export default function Root({ children }) {
-  return (
-    <html>
-      <body>{children}</body>
-    </html>
-  )
-}
