--- test/production/app-dir/revalidate/app/page.js
@@ -1,4 +1,0 @@
-export default async function Page() {
-  const data = Math.random()
-  return <h1>{data}</h1>
-}
