--- test/production/app-dir/revalidate/next.config.js
@@ -1,5 +1,0 @@
-module.exports = {
-  experimental: {
-    appDir: true,
-  },
-}
