--- test/production/app-dir/revalidate/pages/api/revalidate.js
@@ -1,8 +1,0 @@
-export default async function (_req, res) {
-  try {
-    await res.revalidate('/')
-    return res.json({ revalidated: true })
-  } catch (err) {
-    return res.status(500).send('Error')
-  }
-}
